+++ django_api_forms/forms.py
@@ -97,20 +97,23 @@ class BaseForm(object):
 
     def add_error(self, field: Union[str, Tuple], errors: ValidationError):
         if hasattr(errors, 'error_dict'):
-            for items in errors.error_dict.values():
+            for key, items in errors.error_dict.items():
                 for error in items:
                     if isinstance(error, DetailValidationError):
                         error.prepend(field)
                         self.add_error(error.path, error)
                     elif isinstance(error, ValidationError):
-                        self.add_error(field, error)
+                        self.add_error((field, key), error)
         elif not hasattr(errors, 'message') and isinstance(errors.error_list, list):
             for item in errors.error_list:
                 if isinstance(item, DetailValidationError):
                     item.prepend(field)
                     self.add_error(item.path, item)
                 elif isinstance(item, ValidationError):
-                    self.add_error(field, item)
+                    path = field
+                    if hasattr(item, 'path'):
+                        path = (field, ) + item.path
+                    self.add_error(path, item)
         else:
             self._errors.append(
                 DetailValidationError(errors, (field,) if isinstance(field, str) else field)